+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/member/membernews/MemberNews.kt
@@ -1,4 +1,4 @@
-package com.mashup.shorts.domain.member.membernws
+package com.mashup.shorts.domain.member.membernews
 
 import com.mashup.shorts.domain.BaseEntity
 import com.mashup.shorts.domain.member.Member
