+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/member/membernews/MemberNewsRepository.kt
@@ -1,4 +1,4 @@
-package com.mashup.shorts.domain.member.membernws
+package com.mashup.shorts.domain.member.membernews
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
