+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/member/membernewscard/MemberNewsCardClear.kt
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import com.mashup.shorts.common.exception.ShortsBaseException
 import com.mashup.shorts.common.exception.ShortsErrorCode
+import com.mashup.shorts.common.exception.ShortsErrorCode.E404_NOT_FOUND
 import com.mashup.shorts.domain.member.MemberRepository
 import com.mashup.shorts.domain.newscard.NewsCardRepository
 
@@ -18,12 +19,12 @@ class MemberNewsCardClear(
 
     fun clearMemberNewsCard(memberId: Long, newsCardId: Long) {
         val member = memberRepository.findByIdOrNull(memberId) ?: throw ShortsBaseException.from(
-            shortsErrorCode = ShortsErrorCode.E404_MEMBER_NOT_FOUND,
+            shortsErrorCode = E404_NOT_FOUND,
             resultErrorMessage = "${memberId}에 해당하는 유저가 존재하지 않습니다."
         )
 
         val newsCard = newsCardRepository.findByIdOrNull(newsCardId) ?: throw ShortsBaseException.from(
-            shortsErrorCode = ShortsErrorCode.E404_NEWS_CARD_NOT_FOUND,
+            shortsErrorCode = E404_NOT_FOUND,
             resultErrorMessage = "${newsCardId}에 해당하는 뉴스카드가 존재하지 않습니다."
         )
